// src/types/counter.ts
export enum StartStatus {
  start = 'start',
  pause = 'pause',
  stop = 'stop',
}

export enum StartText {
  start = '開始',
  continue = '繼續',
  pause = '暫停',
}

export enum TimeSelectChangeType {
  seconds = 'seconds',
  minutes = 'minutes',
  hour = 'hour',
}

export enum CounterAlertType {
  timesUp = 'times up',
  pleaseSetTime = 'please set time',
}